REF: make Utils dependency-free by moving search_include_directories to Main
